gcp/requirements.txt
# GCP-specific requirements
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.16.4
redis==5.0.1
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1